// src/interfaces/authInterfaces.ts
export interface LoginData{
    name?: string;
    email: string;
    password: string;
}

// Generated by https://quicktype.io
export interface LoginInterface {
    ok:    boolean;
    user:  User;
    token: string;
}

export interface User {
    nombre: string;
    email:  string;
    online: boolean;
    uid:    string;
}